refactor(dashboard): setup GoalAchievement modular structure
Initial setup for GoalAchievement component modularization:
- Created goal-achievement directory structure
- Moved GoalAchievement.tsx to new location
- Added index.ts for clean exports
- Updated imports in dashboard page
- Updated CHANGELOG.md with new refactoring phase

This is the first step in breaking down the 256-line component
into more maintainable, focused pieces.
