+++ docs/refactoring/CHANGELOG.md
@@ -250,3 +250,27 @@
 - Reduced main component complexity from 280 to ~50 lines
 
 ---
+
+## Setup GoalAchievement Refactor Structure
+### Changes
+- Created initial directory structure for GoalAchievement modularization
+- Created directories for card components and utilities
+- Moved GoalAchievement component to dedicated directory
+- Set up index files for clean exports
+
+### Files Added
+- src/components/dashboard/goal-achievement/
+  - cards/          // Goal-specific card components
+    - index.ts      // Card exports
+  - utils/          // Shared utilities
+    - index.ts      // Utility exports
+  - GoalAchievement.tsx
+  - index.ts        // Public exports
+
+### Next Steps
+- Extract BaseGoalCard component
+- Create specific goal card components
+- Extract utility functions for calculations
+- Implement shared configurations
+
+---
