+++ src/components/dashboard/goal-achievement/index.ts
@@ -1,0 +1,1 @@
+export { default as GoalAchievement } from './GoalAchievement';
